+++ frontend/components/ShareButton/spec.jsx
@@ -18,10 +18,18 @@ const mockedParams = {
 };
 
 let mockedIsIOS = true;
-jest.mock('../../helpers/isiOSTheme', () => () => mockedIsIOS);
+jest.mock('@shopgate/pwa-extension-kit', () => ({
+  env: {
+    isIOSTheme: () => mockedIsIOS
+  }
+}));
 
 jest.mock('../../selectors/index', () => ({
   getShareParams: () => mockedParams.mockedShareParams,
+}));
+
+jest.mock('@shopgate/pwa-extension-kit/connectors', () => ({
+  withPageProductId: WrappedComponent => () => <WrappedComponent productId="foo" />,
 }));
 
 /**