+++ frontend/selectors/index.js
@@ -1,12 +1,46 @@
 import { createSelector } from 'reselect';
-import { getCurrentProduct } from '@shopgate/pwa-common-commerce/product/selectors/product';
+import { getProduct } from '@shopgate/pwa-common-commerce/product/selectors/product';
 
+/**
+ * Appends required params for legacy image server.
+ * @param {string} href
+ * @returns {string}
+ */
+const decorateLegacyImageUrl = (href) => {
+  try {
+    const url = new URL(href);
+    url.searchParams.set('w', '880');
+    url.searchParams.set('h', '880');
+
+    return url.toString();
+  } catch (err) {
+    console.error(err);
+    return href;
+  }
+};
+
+/**
+ * Sanitizes image url by detecting legacy urls and appending requided params.
+ * @param {string} url Url.
+ * @returns {string}
+ */
+const sanitizeImageUrl = (url) => {
+  if (!(url && typeof url === 'string' && url.length)) {
+    return '';
+  }
+
+  if (url.startsWith('https://img-cdn.shopgate.com')) {
+    return decorateLegacyImageUrl(url);
+  }
+
+  return url;
+}
 /**
  * Params for share button
  * @return {Array}
  */
 export const getShareParams = createSelector(
-  getCurrentProduct,
+  getProduct,
   (currentProduct) => {
     if (!currentProduct) {
       return null;
@@ -14,7 +48,7 @@ export const getShareParams = createSelector(
 
     const shareParams = {
       title: currentProduct.name,
-      imageURL: currentProduct.featuredImageUrl,
+      imageURL: sanitizeImageUrl(currentProduct.featuredImageUrl),
       deepLink: currentProduct.productUrl,
     };