+++ CHANGELOG.md
@@ -2,6 +2,12 @@
  All notable changes to this project will be documented in this file.
  The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+### [2.0.0] - 2018-01-17
+Major version update.
+### Added
+- Updated styling for Share Buttons to reflect PWA 6 changes
+- Updated connector to receive props for getCurrentProduct selector.
+
 ### [1.0.0] - 2018-10-26
 First version of the extension.
 ### Added